test(dev): split lexical multiple-blocks integration tests
Replace the monolithic lexical-editor-with-multiple-blocks.test.ts with focused files, shared setup helpers, and a chainable nockLexical builder. Adjust directory mock counts so each file passes with its own fresh DB.
