+++ dev/src/tests/fields/lexical-editor-multiple-blocks-build.test.ts
@@ -1,0 +1,536 @@
+import Policies from '../../collections/Policies'
+import { fixture } from './lexical-editor-with-multiple-blocks.fixture'
+import { utilities } from 'payload-crowdin-sync'
+import type { Payload } from 'payload'
+import {
+  injectImageRelationship,
+  setupLexicalMultipleBlocksTest,
+  SNAPSHOT_MEDIA_ID,
+  teardownLexicalMultipleBlocksTest,
+} from './lexical-editor-multiple-blocks-shared.js'
+import { nockLexical } from '../helpers/nock-lexical.js'
+import { assertCrowdinNocksDone, cleanCrowdinNocks } from '../helpers/crowdin-nock.js'
+
+let payload: Payload
+let mediaID: string
+
+
+describe('Lexical editor with multiple blocks - Build', () => {
+  beforeAll(async () => {
+    ;({ payload, mediaID } = await setupLexicalMultipleBlocksTest())
+  })
+  beforeEach(() => {
+    cleanCrowdinNocks()
+  })
+  afterEach(() => {
+    assertCrowdinNocksDone()
+  })
+  afterAll(async () => {
+    await teardownLexicalMultipleBlocksTest(payload)
+  })
+  it('builds a Crowdin HTML object as expected', async () => {
+    nockLexical().directories(3).storages(4).files(4).build()
+    const createdPolicy = await payload.create({
+      collection: 'policies',
+      data: {
+        title: 'Test policy',
+        content: injectImageRelationship(structuredClone(fixture), mediaID),
+      },
+    })
+    const policy = await payload.findByID({
+      collection: 'policies',
+      id: `${createdPolicy.id}`,
+      depth: 0,
+    })
+    expect(
+      injectImageRelationship(
+        structuredClone(
+          utilities.buildCrowdinHtmlObject({
+            doc: policy,
+            fields: Policies.fields,
+          }),
+        ),
+        SNAPSHOT_MEDIA_ID,
+      ),
+    ).toMatchInlineSnapshot(`
+      {
+        "content": {
+          "root": {
+            "children": [
+              {
+                "children": [
+                  {
+                    "detail": 0,
+                    "format": 0,
+                    "mode": "normal",
+                    "style": "",
+                    "text": "Sample content for a Lexical rich text field with multiple blocks.",
+                    "type": "text",
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "type": "paragraph",
+                "version": 1,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "cta",
+                  "id": "65d67d2591c92e447e7472f7",
+                  "link": {
+                    "href": "https://www.npmjs.com/package/payload-crowdin-sync",
+                    "text": "Download payload-crowdin-sync on npm!",
+                    "type": "external",
+                  },
+                  "select": "primary",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "children": [
+                  {
+                    "detail": 0,
+                    "format": 0,
+                    "mode": "normal",
+                    "style": "",
+                    "text": "A bulleted list in-between some blocks consisting of:",
+                    "type": "text",
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "type": "paragraph",
+                "version": 1,
+              },
+              {
+                "children": [
+                  {
+                    "children": [
+                      {
+                        "detail": 0,
+                        "format": 0,
+                        "mode": "normal",
+                        "style": "",
+                        "text": "one bullet list item; and",
+                        "type": "text",
+                        "version": 1,
+                      },
+                    ],
+                    "direction": "ltr",
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 1,
+                    "version": 1,
+                  },
+                  {
+                    "children": [
+                      {
+                        "detail": 0,
+                        "format": 0,
+                        "mode": "normal",
+                        "style": "",
+                        "text": "another!",
+                        "type": "text",
+                        "version": 1,
+                      },
+                    ],
+                    "direction": "ltr",
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 2,
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "listType": "bullet",
+                "start": 1,
+                "tag": "ul",
+                "type": "list",
+                "version": 1,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "highlight",
+                  "color": "green",
+                  "content": {
+                    "root": {
+                      "children": [
+                        {
+                          "children": [
+                            {
+                              "detail": 0,
+                              "format": 0,
+                              "mode": "normal",
+                              "style": "",
+                              "text": "The plugin parses your block configuration for the Lexical rich text editor. It extracts all block values from the rich text field and then treats this config/data combination as a regular \`blocks\` field.",
+                              "type": "text",
+                              "version": 1,
+                            },
+                          ],
+                          "direction": "ltr",
+                          "format": "",
+                          "indent": 0,
+                          "type": "paragraph",
+                          "version": 1,
+                        },
+                        {
+                          "children": [
+                            {
+                              "detail": 0,
+                              "format": 0,
+                              "mode": "normal",
+                              "style": "",
+                              "text": "Markers are placed in the html and this content is restored into the correct place on translation.",
+                              "type": "text",
+                              "version": 1,
+                            },
+                          ],
+                          "direction": "ltr",
+                          "format": "",
+                          "indent": 0,
+                          "type": "paragraph",
+                          "version": 1,
+                        },
+                      ],
+                      "direction": "ltr",
+                      "format": "",
+                      "indent": 0,
+                      "type": "root",
+                      "version": 1,
+                    },
+                  },
+                  "heading": {
+                    "preTitle": "How the plugin handles blocks in the Lexical editor",
+                    "title": "Blocks are extracted into their own fields",
+                  },
+                  "id": "65d67d8191c92e447e7472f8",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "imageText",
+                  "id": "65d67e2291c92e447e7472f9",
+                  "image": "65d67e6a7fb7e9426b3f9f5f",
+                  "title": "Testing a range of fields",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "children": [
+                  {
+                    "children": [],
+                    "direction": null,
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 1,
+                    "version": 1,
+                  },
+                ],
+                "direction": null,
+                "format": "",
+                "indent": 0,
+                "listType": "bullet",
+                "start": 1,
+                "tag": "ul",
+                "type": "list",
+                "version": 1,
+              },
+            ],
+            "direction": "ltr",
+            "format": "",
+            "indent": 0,
+            "type": "root",
+            "version": 1,
+          },
+        },
+      }
+    `)
+  })
+  it('builds a Crowdin JSON object as expected', async () => {
+    nockLexical().directories(2).storages(4).files(4).build()
+    const policy = await payload.create({
+      collection: 'policies',
+      data: {
+        title: 'Test policy',
+        content: injectImageRelationship(structuredClone(fixture), mediaID),
+      },
+    })
+    expect(
+      utilities.buildCrowdinJsonObject({
+        doc: policy,
+        fields: Policies.fields,
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "title": "Test policy",
+      }
+    `)
+  })
+  it('builds a Payload update object as expected', async () => {
+    nockLexical().directories(2).storages(4).files(4).build()
+    const createdPolicy = await payload.create({
+      collection: 'policies',
+      data: {
+        title: 'Test policy',
+        content: injectImageRelationship(structuredClone(fixture), mediaID),
+      },
+    })
+    const policy = await payload.findByID({
+      collection: 'policies',
+      id: `${createdPolicy.id}`,
+      depth: 0,
+    })
+    const crowdinHtmlObject = utilities.buildCrowdinHtmlObject({
+      doc: policy,
+      fields: Policies.fields,
+    })
+    const crowdinJsonObject = utilities.buildCrowdinJsonObject({
+      doc: policy,
+      fields: Policies.fields,
+    })
+    expect(
+      injectImageRelationship(
+        structuredClone(
+          utilities.buildPayloadUpdateObject({
+            crowdinJsonObject,
+            crowdinHtmlObject,
+            fields: Policies.fields,
+            document: policy,
+          }),
+        ),
+        SNAPSHOT_MEDIA_ID,
+      ),
+    ).toMatchInlineSnapshot(`
+      {
+        "content": {
+          "root": {
+            "children": [
+              {
+                "children": [
+                  {
+                    "detail": 0,
+                    "format": 0,
+                    "mode": "normal",
+                    "style": "",
+                    "text": "Sample content for a Lexical rich text field with multiple blocks.",
+                    "type": "text",
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "type": "paragraph",
+                "version": 1,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "cta",
+                  "id": "65d67d2591c92e447e7472f7",
+                  "link": {
+                    "href": "https://www.npmjs.com/package/payload-crowdin-sync",
+                    "text": "Download payload-crowdin-sync on npm!",
+                    "type": "external",
+                  },
+                  "select": "primary",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "children": [
+                  {
+                    "detail": 0,
+                    "format": 0,
+                    "mode": "normal",
+                    "style": "",
+                    "text": "A bulleted list in-between some blocks consisting of:",
+                    "type": "text",
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "type": "paragraph",
+                "version": 1,
+              },
+              {
+                "children": [
+                  {
+                    "children": [
+                      {
+                        "detail": 0,
+                        "format": 0,
+                        "mode": "normal",
+                        "style": "",
+                        "text": "one bullet list item; and",
+                        "type": "text",
+                        "version": 1,
+                      },
+                    ],
+                    "direction": "ltr",
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 1,
+                    "version": 1,
+                  },
+                  {
+                    "children": [
+                      {
+                        "detail": 0,
+                        "format": 0,
+                        "mode": "normal",
+                        "style": "",
+                        "text": "another!",
+                        "type": "text",
+                        "version": 1,
+                      },
+                    ],
+                    "direction": "ltr",
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 2,
+                    "version": 1,
+                  },
+                ],
+                "direction": "ltr",
+                "format": "",
+                "indent": 0,
+                "listType": "bullet",
+                "start": 1,
+                "tag": "ul",
+                "type": "list",
+                "version": 1,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "highlight",
+                  "color": "green",
+                  "content": {
+                    "root": {
+                      "children": [
+                        {
+                          "children": [
+                            {
+                              "detail": 0,
+                              "format": 0,
+                              "mode": "normal",
+                              "style": "",
+                              "text": "The plugin parses your block configuration for the Lexical rich text editor. It extracts all block values from the rich text field and then treats this config/data combination as a regular \`blocks\` field.",
+                              "type": "text",
+                              "version": 1,
+                            },
+                          ],
+                          "direction": "ltr",
+                          "format": "",
+                          "indent": 0,
+                          "type": "paragraph",
+                          "version": 1,
+                        },
+                        {
+                          "children": [
+                            {
+                              "detail": 0,
+                              "format": 0,
+                              "mode": "normal",
+                              "style": "",
+                              "text": "Markers are placed in the html and this content is restored into the correct place on translation.",
+                              "type": "text",
+                              "version": 1,
+                            },
+                          ],
+                          "direction": "ltr",
+                          "format": "",
+                          "indent": 0,
+                          "type": "paragraph",
+                          "version": 1,
+                        },
+                      ],
+                      "direction": "ltr",
+                      "format": "",
+                      "indent": 0,
+                      "type": "root",
+                      "version": 1,
+                    },
+                  },
+                  "heading": {
+                    "preTitle": "How the plugin handles blocks in the Lexical editor",
+                    "title": "Blocks are extracted into their own fields",
+                  },
+                  "id": "65d67d8191c92e447e7472f8",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "fields": {
+                  "blockName": "",
+                  "blockType": "imageText",
+                  "id": "65d67e2291c92e447e7472f9",
+                  "image": "65d67e6a7fb7e9426b3f9f5f",
+                  "title": "Testing a range of fields",
+                },
+                "format": "",
+                "type": "block",
+                "version": 2,
+              },
+              {
+                "children": [
+                  {
+                    "children": [],
+                    "direction": null,
+                    "format": "",
+                    "indent": 0,
+                    "type": "listitem",
+                    "value": 1,
+                    "version": 1,
+                  },
+                ],
+                "direction": null,
+                "format": "",
+                "indent": 0,
+                "listType": "bullet",
+                "start": 1,
+                "tag": "ul",
+                "type": "list",
+                "version": 1,
+              },
+            ],
+            "direction": "ltr",
+            "format": "",
+            "indent": 0,
+            "type": "root",
+            "version": 1,
+          },
+        },
+        "title": "Test policy",
+      }
+    `)
+  })
+})
